+++ examples/cli-example/src/main.playground.ts
@@ -1,6 +1,20 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { initializePlayground, PlaygroundModule } from 'angular-playground';
-import { MyPlaygroundModule } from './my-playground.module';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { ServiceWorkerModule } from '@angular/service-worker';
+import { environment } from './environments/environment';
+import { enableProdMode } from '@angular/core';
 
 initializePlayground('app-root');
-platformBrowserDynamic().bootstrapModule(MyPlaygroundModule);
+
+enableProdMode();
+
+PlaygroundModule
+  .registerRootModules(
+      BrowserAnimationsModule,
+      environment.production ? ServiceWorkerModule.register('/ngsw-worker.js') : []
+  )
+  .enableOverlay();
+
+platformBrowserDynamic().bootstrapModule(PlaygroundModule)
+  .catch(err => console.error(err));