+++ packages/angular-playground/cli/index.ts
@@ -1,3 +1,13 @@
 #!/usr/bin/env node
 import { run } from './src/run';
-run();
+import chalk from 'chalk';
+
+(async () => {
+    try {
+        await run();
+    } catch (err) {
+        console.error(chalk.red(err.message));
+        process.exit(1);
+    }
+})();
+