+++ packages/angular-playground/cli/src/check-errors/verify-sandboxes.ts
@@ -1,15 +1,14 @@
-import * as puppeteer from 'puppeteer';
+import puppeteer = require('puppeteer');
 import { resolve as resolvePath } from 'path';
 import chalk from 'chalk';
-import { copyFileSync } from 'fs';
-import { ErrorReporter, REPORT_TYPE } from './error-reporter';
-import { removeDynamicImports } from './remove-dynamic-imports';
-import { Config } from './configure';
+import { copyFileSync, readFileSync, writeFileSync } from 'fs';
+import { ErrorReporter } from '../error-reporter';
+import { Config } from '../configure';
 
 // Used to tailor the version of headless chromium ran by puppeteer
 const CHROME_ARGS = [ '--disable-gpu', '--no-sandbox' ];
-const SANDBOX_PATH = resolvePath(__dirname, '../../build/shared/sandboxes.js');
-const SANDBOX_DEST = resolvePath(__dirname, '../../sandboxes_modified.js');
+const SANDBOX_PATH = resolvePath(__dirname, '../../../build/src/shared/sandboxes.js');
+const SANDBOX_DEST = resolvePath(__dirname, '../../../sandboxes_modified.js');
 
 export interface ScenarioSummary {
     url: string;
@@ -27,11 +26,11 @@ process.on('unhandledRejection', () => {
     if (browser) browser.close();
 });
 
-export function verifySandboxes(config: Config) {
+export async function verifySandboxes(config: Config) {
     hostUrl = `http://localhost:${config.angularCliPort}`;
     copyFileSync(SANDBOX_PATH, SANDBOX_DEST);
     removeDynamicImports(SANDBOX_DEST);
-    main(config);
+    await main(config);
 }
 
 /////////////////////////////////
@@ -67,7 +66,6 @@ async function main(config: Config) {
  * Creates a Chromium page and navigates to a scenario (URL).
  * If Chromium is not able to connect to the provided page, it will issue a series
  * of retries before it finally fails.
- * @param scenario - Scenario to visit
  */
 async function openScenarioInNewPage(scenario: ScenarioSummary, timeoutAttempts: number) {
     if (timeoutAttempts === 0) {
@@ -126,15 +124,12 @@ function loadSandboxMenuItems(): any[] {
     try {
         return require(SANDBOX_DEST).getSandboxMenuItems();
     } catch (err) {
-        console.log(chalk.red('Failed to load sandbox menu items.'));
-        console.error(err);
-        throw err;
+        throw new Error(`Failed to load sandbox menu items. ${err}`);
     }
 }
 
 /**
  * Callback when Chromium page encounters a console error
- * @param msg - Error message
  */
 async function onConsoleErr(msg: any) {
     if (msg.type === 'error') {
@@ -151,10 +146,17 @@ async function onConsoleErr(msg: any) {
 /**
  * Returns a random value between 1 and the provided length.
  * Note: indexing of keys starts at 1, not 0
- * @param menuItemsLength - Maximum number of items
  */
 function getRandomKey(menuItemsLength: number): number {
     return Math.floor(Math.random() * (menuItemsLength - 1) + 1);
+}
+
+function removeDynamicImports(sandboxPath: string) {
+    const data = readFileSync(sandboxPath, 'utf-8');
+    const dataArray = data.split('\n');
+    const getSandboxIndex = dataArray.findIndex(val => val.includes('getSandbox(path)'));
+    const result = dataArray.slice(0, getSandboxIndex).join('\n');
+    writeFileSync(sandboxPath, result, { encoding: 'utf-8' });
 }
 
 function delay(ms: number) {