+++ packages/angular-playground/cli/src/run.ts
@@ -1,18 +1,22 @@
-import * as program from 'commander';
 import { configure, Config } from './configure';
 import { buildSandboxes } from './build-sandboxes';
 import { startWatch } from './start-watch';
-import { verifySandboxes } from './verify-sandboxes';
-import { findFirstFreePort } from './find-port';
+import { verifySandboxes } from './check-errors/verify-sandboxes';
+import { findFirstFreePort } from './check-errors/find-port';
 import { serveAngularCli } from './serve-angular-cli';
 import { buildAngularCli } from './build-angular-cli';
 
 export async function run() {
     const config: Config = configure(process.argv);
-    await buildSandboxes(config.sourceRoot, config.chunk);
+
+    try {
+        await buildSandboxes(config.sourceRoot, config.chunk);
+    } catch (err) {
+        throw err;
+    }
 
     if (config.buildWithServiceWorkers) {
-        return buildAngularCli(config.angularAppName);
+        return await buildAngularCli(config.angularAppName, config.baseHref);
     }
 
     if (config.verifySandboxes) {
@@ -24,10 +28,18 @@ export async function run() {
     }
 
     if (config.serve || config.verifySandboxes) {
-        serveAngularCli(config);
+        try {
+            await serveAngularCli(config);
+        } catch (err) {
+            throw err;
+        }
     }
 
     if (config.verifySandboxes) {
-        verifySandboxes(config);
+        try {
+            await verifySandboxes(config);
+        } catch (err) {
+            throw err;
+        }
     }
 }