+++ packages/angular-playground/cli/test/files/sandboxes/example1.sandbox.ts
@@ -1,0 +1,13 @@
+export default sandboxOf(ExampleComponent, {
+    imports: [ReactiveFormsModule]
+})
+    .add('Default', {
+        template: `<app-example></app-example>`
+    })
+    .add('With Wrapper', {
+        template: `
+            <div>
+                <app-example></app-example>
+            </div>
+        `
+    });