+++ packages/angular-playground/core/lib/middlewares.ts
@@ -1,0 +1,9 @@
+import { InjectionToken } from '@angular/core';
+
+export interface Middleware {
+    modules?: any[];
+    hostComponent?: any;
+    uiActive: boolean;
+}
+
+export const MIDDLEWARE = new InjectionToken('middleware');