+++ packages/angular-playground/core/src/playground-common.module.ts
@@ -8,9 +8,9 @@ import { UrlService } from './shared/url.service';
 import { AppComponent } from './app.component';
 import { LevenshteinDistance } from './shared/levenshtein-distance';
 import { HighlightSearchMatchPipe } from './shared/highlight-search-match.pipe';
-import { LoaderService } from './shared/loader.service';
 import { LogoComponent } from './logo/logo.component';
 import { PinComponent } from './pin/pin.component';
+import { DrawerComponent } from './drawer/drawer.component';
 
 declare let require: any;
 
@@ -22,7 +22,6 @@ declare let require: any;
     providers: [
         Location,
         { provide: LocationStrategy, useClass: PathLocationStrategy },
-        LoaderService,
         StateService,
         UrlService,
         LevenshteinDistance
@@ -33,7 +32,8 @@ declare let require: any;
         FocusDirective,
         HighlightSearchMatchPipe,
         LogoComponent,
-        PinComponent
+        PinComponent,
+        DrawerComponent
     ],
     exports: [AppComponent]
 })