+++ packages/angular-playground/core/src/scenario/scenario.component.ts
@@ -1,87 +1,137 @@
 import {
-    Component,
-    ViewContainerRef,
-    ChangeDetectionStrategy,
-    NgModule,
-    Injector,
-    Compiler,
-    Input,
-    Inject
+    Component, Input, NgZone, NgModule, OnChanges, OnInit, SimpleChanges, NgModuleRef, Inject, OnDestroy
 } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { Sandbox, Scenario, SelectedSandboxAndScenarioKeys } from '../../lib/app-state';
-import { LoaderService } from '../shared/loader.service';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { SandboxLoader } from '../shared/sandbox-loader';
+import { Scenario, SelectedSandboxAndScenarioKeys } from '../../lib/app-state';
+import { BrowserModule } from '@angular/platform-browser';
+import { Middleware, MIDDLEWARE } from '../../lib/middlewares';
+import { takeUntil } from 'rxjs/operators';
+import { Subject } from 'rxjs/Subject';
 
 @Component({
     selector: 'ap-scenario',
-    template: ``,
-    changeDetection: ChangeDetectionStrategy.OnPush
+    template: `<ng-template></ng-template>`
 })
-export class ScenarioComponent {
-    @Input() set selectedSandboxAndScenarioKeys(selectedSandboxAndScenarioKeys: SelectedSandboxAndScenarioKeys) {
-        this.view.clear();
-        if (selectedSandboxAndScenarioKeys) {
-            this.loaderService.loadSandbox(selectedSandboxAndScenarioKeys.sandboxKey).then(sandbox => {
-                if (sandbox) {
-                    let scenario = sandbox.scenarios.find((s: Scenario) => s.key === selectedSandboxAndScenarioKeys.scenarioKey);
-                    if (scenario) {
-                        this.loadScenario(sandbox, scenario, this.view, this.injector);
-                    }
-                }
-            });
+export class ScenarioComponent implements OnInit, OnChanges, OnDestroy {
+    /**
+     * The selected sandbox and scenario provided from the app dropdown
+     */
+    @Input() selectedSandboxAndScenarioKeys: SelectedSandboxAndScenarioKeys;
+
+    /**
+     * Collection of bootstrapped apps
+     */
+    private activeApps: NgModuleRef<any>[] = [];
+
+    /**
+     * Modules that are applied across every sandbox instance
+     */
+    private activeMiddleware: Middleware;
+
+    /**
+     * Unsubscribe all subscriptions on component destroy
+     */
+    private onDestroy = new Subject<void>();
+
+    constructor(private zone: NgZone, @Inject(MIDDLEWARE) private middleware) {
+    }
+
+    ngOnInit() {
+        this.middleware
+            .pipe(takeUntil(this.onDestroy))
+            .subscribe(middlewares => this.activeMiddleware = middlewares);
+
+        if (this.selectedSandboxAndScenarioKeys) {
+            this.bootstrapSandbox(this.selectedSandboxAndScenarioKeys);
         }
     }
 
-    constructor(private loaderService: LoaderService,
-        private compiler: Compiler,
-        private injector: Injector,
-        private view: ViewContainerRef) {
+    ngOnChanges(changes: SimpleChanges) {
+        if (changes.selectedSandboxAndScenarioKeys) {
+            this.bootstrapSandbox(changes.selectedSandboxAndScenarioKeys.currentValue)
+        }
     }
 
-    private loadScenario(sandbox: Sandbox, scenario: Scenario, location: ViewContainerRef, injector: Injector) {
-        let hostComponent = this.hostComponent(scenario);
-        let hostModule = this.hostModule(sandbox, hostComponent);
-        this.compiler.clearCache();
-        let moduleFactory = this.compiler.compileModuleSync(hostModule);
-        let moduleRef = moduleFactory.create(injector);
-        let componentFactory = moduleRef.componentFactoryResolver.resolveComponentFactory(hostComponent);
-        location.createComponent(componentFactory, location.length, moduleRef.injector);
+    ngOnDestroy() {
+        this.onDestroy.next();
     }
 
-    private hostComponent(scenario: Scenario) {
-        @Component({
-            selector: 'host-component',
-            template: scenario.template,
-            styles: scenario.styles,
-            providers: scenario.providers
-        })
-        class HostComponent {
+    /**
+     * Bootstrap a new Angular application with the sandbox's required dependencies
+     */
+    private bootstrapSandbox(selectedSandboxAndScenarioKeys: SelectedSandboxAndScenarioKeys) {
+        SandboxLoader.loadSandbox(selectedSandboxAndScenarioKeys.sandboxKey).then(sandbox => {
+            if (sandbox) {
+                const scenario = sandbox.scenarios
+                    .find((s: Scenario) => s.key === selectedSandboxAndScenarioKeys.scenarioKey);
+
+                if (scenario) {
+                    if (this.activeApps.length > 0) {
+                        const app = this.activeApps.pop();
+                        app.destroy();
+                    }
+
+                    // Don't bootstrap a new Angular application within an existing zone
+                    this.zone.runOutsideAngular(() => {
+                        const module = this.createModule(sandbox, scenario);
+                        platformBrowserDynamic().bootstrapModule(module)
+                            .then(app => this.activeApps.push(app))
+                            .catch(err => console.error(err));
+                    });
+                }
+            }
+        });
+    }
+
+    /**
+     * Create a module containing the dependencies of a sandbox
+     */
+    private createModule(sandboxMeta, scenario) {
+        const hostComp = this.createComponent(scenario);
+
+        class DynamicModule {
+            ngDoBootstrap(app) {
+                const hostEl = document.querySelector('playground-host');
+                if (!hostEl) {
+                    const compEl = document.createElement('playground-host');
+                    document.body.appendChild(compEl);
+                }
+                app.bootstrap(hostComp);
+            }
+        }
+
+        return NgModule({
+            imports: [
+                BrowserModule,
+                ...sandboxMeta.imports,
+                ...this.activeMiddleware.modules
+            ],
+            declarations: [
+                hostComp,
+                sandboxMeta.declareComponent ? sandboxMeta.type : [],
+                ...sandboxMeta.declarations
+            ],
+            providers: [...sandboxMeta.providers],
+            entryComponents: [hostComp]
+        })(DynamicModule);
+    }
+
+    /**
+     * Construct a component to serve as the host for the provided scenario
+     */
+    private createComponent(scenario: Scenario) {
+        class DynamicComponent {
             constructor() {
                 Object.assign(this, scenario.context);
             }
         }
-        return HostComponent;
-    }
 
-    private hostModule(sandbox: Sandbox, hostComponent: any) {
-        let { imports, type, declarations, providers } = sandbox;
-        @NgModule({
-            imports: [
-                CommonModule,
-                imports ? imports : []
-            ],
-            declarations: [
-                ...(sandbox.declareComponent ? [type] : []),
-                hostComponent,
-                declarations ? declarations : []
-            ],
-            providers: providers ? providers : [],
-            entryComponents: [
-                hostComponent
-            ]
-        })
-        class HostModule {
-        }
-        return HostModule;
+        return Component({
+            selector: 'playground-host',
+            template: scenario.template,
+            styles: scenario.styles,
+            providers: scenario.providers
+        })(DynamicComponent);
     }
 }