+++ packages/angular-playground/core/src/shared/url.service.ts
@@ -1,7 +1,7 @@
-import { Inject, Injectable } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { Location } from '@angular/common';
 import { SandboxMenuItem } from '../../lib/app-state';
-import { LoaderService } from './loader.service';
+import { SandboxLoader } from './sandbox-loader';
 
 @Injectable()
 export class UrlService {
@@ -18,9 +18,8 @@ export class UrlService {
         return this._select;
     }
 
-    constructor(private loaderService: LoaderService,
-        private location: Location) {
-        this.sandboxMenuItems = this.loaderService.getSandboxMenuItems();
+    constructor(private location: Location) {
+        this.sandboxMenuItems = SandboxLoader.getSandboxMenuItems();
         let urlPath = location.path();
         this._embed = /[?|&]embed=1/.exec(urlPath) !== null;
         this._select = this.parse('scenario', this.sandboxMenuItems, urlPath);